refactor : 숫자생성 도메인 객체를 생성하는 방식을 변경한다
숫자 생성 도메인에서 전략을 통해 숫자를 생성하는 것이 아닌, Randoms 함수를 사용하도록 변경한다.
- 랜덤한 숫자를 생성하는 Randoms 패키지를 외부호출(infra) 라고 생각하여 전략을 통해 구현하려 함
- 하지만, 해당 Randoms 패키지는 랜덤숫자생성과 매우 밀접하기에 infra 로 생각하지 않고 도메인 로직으로 생각함
