+++ README.md
@@ -65,7 +65,6 @@ june : ----
 - [x] 자동차들 중, 가장 큰 위치값이 몇인지 판별한다.
 - [x] 자동차들 중, 최종우승자를 반환한다.
 - [x] 숫자생성 도메인 객체를 생성한다.
-- [ ] 0-9 사이의 랜덤값을 생성하는 전략을 구현한다.
 - [ ] 랜덤값을 반환하는 서비스를 구현한다.
 - [ ] 자동차가 랜덤값을 받는 전략을 구현한다.
 - [ ] 자동차들이 랜덤값을 받아 움직이도록 구현한다.
