--- src/main/java/racingcar/generator/random/domain/external/NumberGenerateStrategy.java
@@ -1,6 +1,0 @@
-package racingcar.generator.random.domain.external;
-
-public interface NumberGenerateStrategy {
-
-    int generateNumberInRangeInclusively(int startInclusive, int endInclusive);
-}
